Adding the technique T1219 to the framework.
Utilizing QuickSupport from Teamviewer we achieve persistence by running it without installing and dumping ID and Password from the process memory.
