+++ Techniques/T1219/T1219.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30406.217
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "T1219", "T1219.csproj", "{272501A9-AE90-4807-83F6-7763D0CC9768}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{272501A9-AE90-4807-83F6-7763D0CC9768}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{272501A9-AE90-4807-83F6-7763D0CC9768}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{272501A9-AE90-4807-83F6-7763D0CC9768}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{272501A9-AE90-4807-83F6-7763D0CC9768}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7F73D130-6DE9-4CE0-A405-E22E9147FA19}
+	EndGlobalSection
+EndGlobal
